simulator_test: Move LimitMalloc test; skip separate test on --config=asan
