+++ systems/analysis/BUILD.bazel
@@ -438,7 +438,6 @@ drake_cc_googletest(
         ":simulator",
         "//common/test_utilities:expect_throws_message",
         "//common/test_utilities:is_dynamic_castable",
-        "//common/test_utilities:limit_malloc",
         "//systems/analysis/test_utilities:controlled_spring_mass_system",
         "//systems/analysis/test_utilities:logistic_system",
         "//systems/analysis/test_utilities:my_spring_mass_system",
@@ -446,6 +445,18 @@ drake_cc_googletest(
         "//systems/primitives:constant_vector_source",
         "//systems/primitives:integrator",
         "//systems/primitives:signal_logger",
+    ],
+)
+
+drake_cc_googletest(
+    name = "simulator_limit_malloc_test",
+    # TODO(#14901): This usage of LimitMalloc does not play nicely with Asan.
+    tags = ["no_asan"],
+    deps = [
+        ":simulator",
+        "//common/test_utilities:limit_malloc",
+        "//systems/framework:diagram_builder",
+        "//systems/framework:leaf_system",
     ],
 )
